+++ CHANGELOG.md
@@ -36,6 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added more examples.
 - xDnsRecordMx
   - Added new resource to manage MX records
+- xDnsServerZoneAging
+  - Added integration tests ([issue #176](https://github.com/dsccommunity/xDnsServer/issues/176)).
 
 ### Changed